+++ common/getziti/install_ziti_edge_tunnel.go
@@ -13,7 +13,7 @@ func InstallZitiEdgeTunnel(targetVersion, targetOS, targetArch, binDir string, v
 	if targetVersion != "" {
 		newVersion = semver.MustParse(strings.TrimPrefix(targetVersion, "v"))
 	} else {
-		v, err := GetLatestGitHubReleaseVersion(c.ZITI_EDGE_TUNNEL_GITHUB, verbose)
+		v, err := GetLatestGitHubReleaseVersion(c.OpenZitiOrg, c.ZITI_EDGE_TUNNEL_GITHUB, verbose)
 		if err != nil {
 			return err
 		}
@@ -22,5 +22,5 @@ func InstallZitiEdgeTunnel(targetVersion, targetOS, targetArch, binDir string, v
 
 	fmt.Println("Attempting to install '" + c.ZITI_EDGE_TUNNEL + "' version: " + newVersion.String())
 	return FindVersionAndInstallGitHubRelease(
-		c.ZITI_EDGE_TUNNEL, c.ZITI_EDGE_TUNNEL_GITHUB, targetOS, targetArch, binDir, newVersion.String(), verbose)
+		c.OpenZitiOrg, c.ZITI_EDGE_TUNNEL, c.ZITI_EDGE_TUNNEL_GITHUB, targetOS, targetArch, binDir, newVersion.String(), verbose)
 }