+++ zititest/models/test_resources/terraform/vpc/main.tf
@@ -64,6 +64,18 @@ resource "aws_security_group" "fablab" {
   vpc_id = aws_vpc.fablab.id
 
   ingress {
+    from_port   = 443
+    to_port     = 443
+    protocol    = "tcp"
+    cidr_blocks = ["0.0.0.0/0"]
+  }
+  ingress {
+    from_port   = 443
+    to_port     = 443
+    protocol    = "udp"
+    cidr_blocks = ["0.0.0.0/0"]
+  }
+  ingress {
     from_port   = 10000
     to_port     = 10000
     protocol    = "tcp"